# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "nkuwiki API",
        "description": "\u5357\u5f00\u767e\u79d1\u77e5\u8bc6\u5e73\u53f0API\u670d\u52a1",
        "version": "1.0.0"
    },
    "paths": {
...
        "/wxapp/search": {
            "post": {
                "tags": [
                    "wxapp"
                ],
                "summary": "Search",
                "description": "\u6267\u884c\u641c\u7d22\u67e5\u8be2\uff0c\u8fd4\u56de\u6700\u76f8\u5173\u7684\u7ed3\u679c",
                "operationId": "search_wxapp_search_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SearchQuery"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SearchResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/wxapp/index-document": {
            "post": {
                "tags": [
                    "wxapp"
                ],
                "summary": "Index Document",
                "description": "\u4e3a\u5fae\u4fe1\u4e91\u5b58\u50a8\u4e2d\u7684\u6587\u4ef6\u5efa\u7acb\u7d22\u5f15\uff0c\u652f\u6301\u591a\u79cd\u6587\u6863\u683c\u5f0f\n\n\u6b64API\u63a5\u6536\u5fae\u4fe1\u4e91\u5b58\u50a8\u7684fileID\uff0c\u4ece\u4e2d\u68c0\u7d22\u6587\u4ef6\u5185\u5bb9\uff0c\u7136\u540e\u5efa\u7acb\u68c0\u7d22\u7d22\u5f15\n\u652f\u6301\u7684\u6587\u4ef6\u683c\u5f0f: PDF, DOCX, TXT, MARKDOWN \u7b49",
                "operationId": "index_document_wxapp_index_document_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentIndexRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentIndexResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/wxapp/search_post": {
            "post": {
                "tags": [
                    "wxapp"
                ],
                "summary": "\u9ad8\u7ea7\u641c\u7d22\u5e16\u5b50",
                "description": "\u9ad8\u7ea7\u641c\u7d22\u5e16\u5b50\u529f\u80fd\uff0c\u652f\u6301\u591a\u6761\u4ef6\u7075\u6d3b\u641c\u7d22\uff0c\u5305\u62ec\u5173\u952e\u8bcd\u3001\u6807\u9898\u3001\u5185\u5bb9\u3001\u53d1\u5e16\u4eba\u7b49",
                "operationId": "search_post_wxapp_search_post_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SearchPostRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Search Post Wxapp Search Post Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DocumentIndexRequest": {
                "properties": {
                    "file_id": {
                        "type": "string",
                        "title": "File Id",
                        "description": "\u5fae\u4fe1\u4e91\u5b58\u50a8\u7684\u6587\u4ef6ID"
                    },
                    "file_name": {
                        "type": "string",
                        "title": "File Name",
                        "description": "\u6587\u4ef6\u540d\u79f0"
                    },
                    "openid": {
                        "type": "string",
                        "title": "Openid",
                        "description": "\u7528\u6237openid"
                    },
                    "custom_metadata": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Custom Metadata",
                        "description": "\u81ea\u5b9a\u4e49\u5143\u6570\u636e"
                    }
                },
                "type": "object",
                "required": [
                    "file_id",
                    "file_name",
                    "openid"
                ],
                "title": "DocumentIndexRequest"
            },
            "DocumentIndexResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "title": "Success"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "document_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Document Id"
                    },
                    "suggested_keywords": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Suggested Keywords"
                    }
                },
                "type": "object",
                "required": [
                    "success",
                    "message"
                ],
                "title": "DocumentIndexResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "SearchPostRequest": {
                "properties": {
                    "keyword": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Keyword",
                        "description": "\u5173\u952e\u8bcd\uff0c\u641c\u7d22\u6807\u9898\u548c\u5185\u5bb9"
                    },
                    "title": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Title",
                        "description": "\u6807\u9898\u5173\u952e\u8bcd"
                    },
                    "content": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Content",
                        "description": "\u5185\u5bb9\u5173\u952e\u8bcd"
                    },
                    "openid": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Openid",
                        "description": "\u6309\u53d1\u5e16\u7528\u6237openid\u7b5b\u9009"
                    },
                    "nick_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Nick Name",
                        "description": "\u6309\u7528\u6237\u6635\u79f0\u7b5b\u9009"
                    },
                    "tags": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tags",
                        "description": "\u6309\u6807\u7b7e\u7b5b\u9009\uff0c\u652f\u6301\u591a\u4e2a\u6807\u7b7e"
                    },
                    "category_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Category Id",
                        "description": "\u6309\u5206\u7c7bID\u7b5b\u9009"
                    },
                    "start_time": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Start Time",
                        "description": "\u5f00\u59cb\u65f6\u95f4\uff0c\u683c\u5f0f\uff1aYYYY-MM-DD HH:MM:SS"
                    },
                    "end_time": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "End Time",
                        "description": "\u7ed3\u675f\u65f6\u95f4\uff0c\u683c\u5f0f\uff1aYYYY-MM-DD HH:MM:SS"
                    },
                    "status": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Status",
                        "description": "\u5e16\u5b50\u72b6\u6001: 1-\u6b63\u5e38, 0-\u7981\u7528",
                        "default": 1
                    },
                    "include_deleted": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Include Deleted",
                        "description": "\u662f\u5426\u5305\u542b\u5df2\u5220\u9664\u7684\u5e16\u5b50",
                        "default": false
                    },
                    "sort_by": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Sort By",
                        "description": "\u6392\u5e8f\u65b9\u5f0f\uff0c\u652f\u6301create_time/update_time/like_count/comment_count ASC/DESC",
                        "default": "update_time DESC"
                    },
                    "page": {
                        "type": "integer",
                        "minimum": 1.0,
                        "title": "Page",
                        "description": "\u9875\u7801",
                        "default": 1
                    },
                    "page_size": {
                        "type": "integer",
                        "maximum": 100.0,
                        "minimum": 1.0,
                        "title": "Page Size",
                        "description": "\u6bcf\u9875\u8bb0\u5f55\u6570",
                        "default": 20
                    }
                },
                "type": "object",
                "title": "SearchPostRequest",
                "description": "\u5e16\u5b50\u641c\u7d22\u8bf7\u6c42"
            },
            "SearchQuery": {
                "properties": {
                    "query": {
                        "type": "string",
                        "title": "Query",
                        "description": "\u641c\u7d22\u67e5\u8be2\u8bcd"
                    },
                    "limit": {
                        "type": "integer",
                        "title": "Limit",
                        "description": "\u8fd4\u56de\u7ed3\u679c\u6570\u91cf\u9650\u5236",
                        "default": 10
                    },
                    "offset": {
                        "type": "integer",
                        "title": "Offset",
                        "description": "\u8d77\u59cb\u504f\u79fb\u91cf",
                        "default": 0
                    },
                    "filters": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Filters",
                        "description": "\u8fc7\u6ee4\u6761\u4ef6"
                    }
                },
                "type": "object",
                "required": [
                    "query"
                ],
                "title": "SearchQuery"
            },
            "SearchResponse": {
                "properties": {
                    "results": {
                        "items": {
                            "$ref": "#/components/schemas/SearchResult"
                        },
                        "type": "array",
                        "title": "Results"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    },
                    "query": {
                        "type": "string",
                        "title": "Query"
                    }
                },
                "type": "object",
                "required": [
                    "results",
                    "total",
                    "query"
                ],
                "title": "SearchResponse"
            },
            "SearchResult": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content"
                    },
                    "score": {
                        "type": "number",
                        "title": "Score"
                    },
                    "source": {
                        "type": "string",
                        "title": "Source"
                    },
                    "url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Url"
                    },
                    "metadata": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Metadata"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "title",
                    "content",
                    "score",
                    "source"
                ],
                "title": "SearchResult"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}